src/python/core/pyproject.toml
[tool.poetry]
name = "semio"
version = "0.1.0"
description = "Semio allows designers to explicitly formalize design intent. This allows reuse, to split and parallelize work, producing more flexible systems."
authors = ["Ueli Saluz <semio@posteo.org>"]
license = "AGPL"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.23.5"
pytest = "^7.2.0"
networkx = "^2.8.8"
mathutils = "^3.3.0"
multipledispatch = "^0.6.0"
pytest-lazy-fixture = "^0.6.3"
pydantic = "^1.10.2"
dacite = "^1.6.0"
grpcio = "^1.51.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
